+++ src/flick/tasks.py
@@ -1,14 +1,9 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from user.models import Profile
-
 from celery import shared_task
 from imdb import IMDb
-from notification.models import Notification
 from provider.models import Provider
-from push_notifications.models import APNSDevice
-from push_notifications.models import GCMDevice
 from show.animelist import flickanimelist
 from show.models import Show
 from show.tmdb import flicktmdb
@@ -129,30 +124,3 @@ def populate_show_details(show_id):
         imdb_rating = imdb_api.get_movie(imdb_id).get("rating")
         show.imdb_rating = imdb_rating
     show.save()
-
-
-@shared_task
-def create_incoming_friend_request_accepted_for_to_user(from_user, to_user):
-    from_profile = Profile.objects.get(user=from_user)
-    to_profile = Profile.objects.get(user=to_user)
-    notif = Notification()
-    notif.notif_type = "incoming_friend_request_accepted"
-    notif.from_user = from_profile
-    notif.to_user = to_profile
-    notif.save()
-
-
-@shared_task
-def create_outgoing_friend_request_accepted_for_from_user(from_user, to_user):
-    from_profile = Profile.objects.get(user=from_user)
-    to_profile = Profile.objects.get(user=to_user)
-    notif = Notification()
-    notif.notif_type = "outgoing_friend_request_accepted"
-    notif.from_user = to_profile
-    notif.to_user = from_profile
-    notif.save()
-    ios_devices = APNSDevice.objects.filter(user=from_user, active=True)
-    android_devices = GCMDevice.objects.filter(user=from_user, active=True)
-    message_body = f"({from_user.username}): {to_user.first_name} (@{to_user.username}) accepted your friend request."
-    ios_devices.send_message(message={"body": message_body})
-    android_devices.send_message(message={"body": message_body})
